requirements.txt
matplotlib==3.0.3
numpy==1.16.5
opencv-python==4.0.0.21
pandas==0.24.2
plyfile==0.7
PyYAML==5.1
scipy==1.4.1
scikit-image==0.15.0
seaborn==0.9.0
sklearn==0.0
trimesh==3.9.21
tensorboard==2.5.0
tensorflow-gpu==2.3.0
tensorflow-addons==0.13.0
tensorflow-graphics-gpu
wget==3.2